+++ src/main/java/ru/vyarus/guice/persist/orient/db/pool/graph/GraphPool.java
@@ -1,7 +1,7 @@
 package ru.vyarus.guice.persist.orient.db.pool.graph;
 
 import com.orientechnologies.orient.core.db.ODatabaseDocumentInternal;
-import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx;
+import com.orientechnologies.orient.core.db.OrientDB;
 import com.orientechnologies.orient.core.tx.OTransaction;
 import com.tinkerpop.blueprints.impls.orient.OrientBaseGraph;
 import com.tinkerpop.blueprints.impls.orient.OrientGraph;
@@ -9,12 +9,13 @@ import com.tinkerpop.blueprints.impls.orient.OrientGraphNoTx;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.vyarus.guice.persist.orient.db.DbType;
-import ru.vyarus.guice.persist.orient.db.pool.PoolManager;
 import ru.vyarus.guice.persist.orient.db.pool.DocumentPool;
+import ru.vyarus.guice.persist.orient.db.pool.PoolManager;
 import ru.vyarus.guice.persist.orient.db.transaction.TransactionManager;
 import ru.vyarus.guice.persist.orient.db.user.UserManager;
 
 import javax.inject.Inject;
+import javax.inject.Provider;
 
 /**
  * Graph pool implementation. Provides base graph connection (see additional providers for specific graph connections).
@@ -27,24 +28,30 @@ public class GraphPool implements PoolManager<OrientBaseGraph> {
     private final Logger logger = LoggerFactory.getLogger(GraphPool.class);
 
     private final ThreadLocal<OrientBaseGraph> transaction = new ThreadLocal<OrientBaseGraph>();
+    private final Provider<OrientDB> orientDB;
     private final TransactionManager transactionManager;
     private final DocumentPool documentPool;
     private final UserManager userManager;
 
     @Inject
-    public GraphPool(final TransactionManager transactionManager, final DocumentPool documentPool,
+    public GraphPool(final Provider<OrientDB> orientDB,
+                     final TransactionManager transactionManager,
+                     final DocumentPool documentPool,
                      final UserManager userManager) {
+        this.orientDB = orientDB;
         this.transactionManager = transactionManager;
         this.documentPool = documentPool;
         this.userManager = userManager;
     }
 
     @Override
-    public void start(final String uri) {
+    public void start(final String database) {
         // test connection and let orient configure database
-        new OrientGraph(new ODatabaseDocumentTx(uri)
-                .<ODatabaseDocumentTx>open(userManager.getUser(), userManager.getPassword())).getRawGraph().close();
-        logger.debug("Pool {} started for '{}'", getType(), uri);
+        new OrientGraph(
+                (ODatabaseDocumentInternal) orientDB.get()
+                        .open(database, userManager.getUser(), userManager.getPassword()))
+                .getRawGraph().close();
+        logger.debug("Pool {} started for database '{}'", getType(), database);
     }
 
     @Override