+++ src/test/groovy/ru/vyarus/guice/persist/orient/db/ExternalTxTest.groovy
@@ -1,7 +1,7 @@
 package ru.vyarus.guice.persist.orient.db
 
 import com.orientechnologies.orient.core.db.ODatabaseRecordThreadLocal
-import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx
+import com.orientechnologies.orient.core.db.OrientDB
 import com.orientechnologies.orient.core.exception.ODatabaseException
 import com.orientechnologies.orient.core.tx.OTransaction
 import ru.vyarus.guice.persist.orient.AbstractTest
@@ -11,6 +11,9 @@ import ru.vyarus.guice.persist.orient.support.Config
 import ru.vyarus.guice.persist.orient.support.modules.PackageSchemeModule
 import spock.guice.UseModules
 
+import javax.inject.Inject
+import javax.inject.Provider
+
 /**
  * @author Vyacheslav Rusakov
  * @since 21.10.2017
@@ -18,13 +21,18 @@ import spock.guice.UseModules
 @UseModules(PackageSchemeModule)
 class ExternalTxTest extends AbstractTest {
 
+    @Inject
+    Provider<OrientDB> orientDB;
+
+
     def "Check external connection usage"() {
 
         setup: "manually creating connection"
-        def extDb = new ODatabaseDocumentTx(Config.DB).open(Config.USER, Config.PASS)
+        // using same orientdb object to simplify initialization
+        def extDb = orientDB.get().open(Config.getDbName(), Config.USER, Config.PASS);
 
         when: "work in external unit of work"
-        context.doInTransaction(TxConfig.external(), {db ->
+        context.doInTransaction(TxConfig.external(), { db ->
             assert context.transactionManager.externalTransaction
             assert context.transactionManager.activeTransactionType == OTransaction.TXTYPE.NOTX
             assert extDb == db.underlying // same underlining instance
@@ -34,16 +42,16 @@ class ExternalTxTest extends AbstractTest {
         !extDb.isClosed()
 
         cleanup:
-        extDb.close()
+        extDb?.close()
     }
 
     def "Check external connection rollback"() {
 
         setup: "manually creating connection"
-        def extDb = new ODatabaseDocumentTx(Config.DB).open(Config.USER, Config.PASS)
+        def extDb = orientDB.get().open(Config.getDbName(), Config.USER, Config.PASS);
 
         when: "throw exception during unit of work"
-        context.doInTransaction(TxConfig.external(), {db ->
+        context.doInTransaction(TxConfig.external(), { db ->
             throw new IllegalStateException("UPS")
         } as SpecificTxAction)
         then: "external connection wasn't rolled back"
@@ -52,7 +60,7 @@ class ExternalTxTest extends AbstractTest {
 
 
         cleanup:
-        extDb.close()
+        extDb?.close()
     }
 
     def "Check external unit did not start without thread bound connection"() {
@@ -67,11 +75,11 @@ class ExternalTxTest extends AbstractTest {
     def "Check optimistic tx recognition"() {
 
         setup: "manually creating connection"
-        def extDb = new ODatabaseDocumentTx(Config.DB).open(Config.USER, Config.PASS)
+        def extDb = orientDB.get().open(Config.getDbName(), Config.USER, Config.PASS);
         extDb.begin(OTransaction.TXTYPE.OPTIMISTIC)
 
         when: "work in external unit of work"
-        context.doInTransaction(TxConfig.external(), {db ->
+        context.doInTransaction(TxConfig.external(), { db ->
             assert context.transactionManager.externalTransaction
             assert context.transactionManager.activeTransactionType == OTransaction.TXTYPE.OPTIMISTIC
         } as SpecificTxAction)
@@ -79,20 +87,22 @@ class ExternalTxTest extends AbstractTest {
         ODatabaseRecordThreadLocal.instance().isDefined()
 
         cleanup:
-        extDb.close()
+        extDb?.close()
     }
 
     def "Check external connection close in between"() {
 
         setup: "manually creating connection"
-        new ODatabaseDocumentTx(Config.DB).open(Config.USER, Config.PASS)
+        def extDb = orientDB.get().open(Config.getDbName(), Config.USER, Config.PASS);
 
         when: "work in external unit of work"
-        context.doInTransaction(TxConfig.external(), {db ->
+        context.doInTransaction(TxConfig.external(), { db ->
             db.close()
         } as SpecificTxAction)
         then: "close detected"
         thrown(IllegalStateException)
 
+        cleanup:
+        extDb?.close()
     }
 }