+++ src/test/groovy/ru/vyarus/guice/persist/orient/db/pool/support/pool/NoTypePool.groovy
@@ -13,7 +13,7 @@ import ru.vyarus.guice.persist.orient.db.pool.PoolManager
 class NoTypePool implements PoolManager<ODatabaseDocument> {
 
     @Override
-    void start(String uri) {
+    void start(String database) {
 
     }